move non windows additional notes to new file
